+++ Indicators/Stochastic/Stoch.Models.cs
@@ -3,12 +3,12 @@
 
     public class StochResult : ResultBase
     {
-        public float? Oscillator { get; set; }
-        public float? Signal { get; set; }
+        public decimal? Oscillator { get; set; }
+        public decimal? Signal { get; set; }
         public bool? IsIncreasing { get; set; }
 
         // internal use only
-        internal float? Smooth { get; set; }
+        internal decimal? Smooth { get; set; }
     }
 
 }
